refactor: move CSV-related annotation & interface to annotations module
Fixes #700
