+++ annotations/src/main/java/io/quarkiverse/operatorsdk/annotations/SharedCSVMetadata.java
@@ -1,0 +1,4 @@
+package io.quarkiverse.operatorsdk.annotations;
+
+public interface SharedCSVMetadata {
+}
