+++ bundle-generator/deployment/pom.xml
@@ -37,6 +37,10 @@
             <artifactId>quarkus-junit5-internal</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>io.quarkiverse.operatorsdk</groupId>
+            <artifactId>quarkus-operator-sdk-annotations</artifactId>
+        </dependency>
     </dependencies>
 
     <build>
