+++ bundle-generator/deployment/src/test/java/io/quarkiverse/operatorsdk/bundle/sources/BReconciler.java
@@ -4,7 +4,7 @@ import io.fabric8.kubernetes.api.model.Service;
 import io.javaoperatorsdk.operator.api.reconciler.Context;
 import io.javaoperatorsdk.operator.api.reconciler.Reconciler;
 import io.javaoperatorsdk.operator.api.reconciler.UpdateControl;
-import io.quarkiverse.operatorsdk.bundle.runtime.CSVMetadata;
+import io.quarkiverse.operatorsdk.annotations.CSVMetadata;
 
 @CSVMetadata(name = "shared", version = "0.0.2")
 public class BReconciler implements Reconciler<Service> {
