+++ bundle-generator/deployment/src/test/java/io/quarkiverse/operatorsdk/bundle/sources/DuplicatedBundleNameWithoutSharedCSVMetadata1.java
@@ -4,7 +4,7 @@ import io.fabric8.openshift.api.model.Role;
 import io.javaoperatorsdk.operator.api.reconciler.Context;
 import io.javaoperatorsdk.operator.api.reconciler.Reconciler;
 import io.javaoperatorsdk.operator.api.reconciler.UpdateControl;
-import io.quarkiverse.operatorsdk.bundle.runtime.CSVMetadata;
+import io.quarkiverse.operatorsdk.annotations.CSVMetadata;
 
 @CSVMetadata(name = "illegal")
 public class DuplicatedBundleNameWithoutSharedCSVMetadata1 implements Reconciler<Role> {
