+++ bundle-generator/deployment/src/test/java/io/quarkiverse/operatorsdk/bundle/sources/External.java
@@ -3,7 +3,7 @@ package io.quarkiverse.operatorsdk.bundle.sources;
 import io.fabric8.kubernetes.client.CustomResource;
 import io.fabric8.kubernetes.model.annotation.Group;
 import io.fabric8.kubernetes.model.annotation.Version;
-import io.quarkiverse.operatorsdk.bundle.runtime.CSVMetadata;
+import io.quarkiverse.operatorsdk.annotations.CSVMetadata;
 
 @Group("halkyon.io")
 @Version("v1alpha1")
