+++ bundle-generator/deployment/src/test/java/io/quarkiverse/operatorsdk/bundle/sources/FirstReconciler.java
@@ -3,7 +3,7 @@ package io.quarkiverse.operatorsdk.bundle.sources;
 import io.javaoperatorsdk.operator.api.reconciler.Context;
 import io.javaoperatorsdk.operator.api.reconciler.Reconciler;
 import io.javaoperatorsdk.operator.api.reconciler.UpdateControl;
-import io.quarkiverse.operatorsdk.bundle.runtime.CSVMetadata;
+import io.quarkiverse.operatorsdk.annotations.CSVMetadata;
 
 @CSVMetadata(name = "first-operator", version = FirstReconciler.VERSION)
 public class FirstReconciler implements Reconciler<First> {
