+++ bundle-generator/deployment/src/test/java/io/quarkiverse/operatorsdk/bundle/sources/Third.java
@@ -4,7 +4,7 @@ import io.fabric8.kubernetes.api.model.Namespaced;
 import io.fabric8.kubernetes.client.CustomResource;
 import io.fabric8.kubernetes.model.annotation.Group;
 import io.fabric8.kubernetes.model.annotation.Version;
-import io.quarkiverse.operatorsdk.bundle.runtime.CSVMetadata;
+import io.quarkiverse.operatorsdk.annotations.CSVMetadata;
 
 @Group("samples.javaoperatorsdk.io")
 @Version("v1alpha1")
