+++ bundle-generator/runtime/pom.xml
@@ -24,6 +24,10 @@
             <groupId>io.quarkus</groupId>
             <artifactId>quarkus-kubernetes</artifactId>
         </dependency>
+        <dependency>
+            <groupId>io.quarkiverse.operatorsdk</groupId>
+            <artifactId>quarkus-operator-sdk-annotations</artifactId>
+        </dependency>
     </dependencies>
 
     <build>
