+++ bundle-generator/runtime/src/main/java/io/quarkiverse/operatorsdk/bundle/runtime/SharedCSVMetadata.java
@@ -1,4 +1,9 @@
 package io.quarkiverse.operatorsdk.bundle.runtime;
 
+@SuppressWarnings("unused")
+@Deprecated(since = "6.3.1", forRemoval = true)
+/**
+ * @deprecated Use {@link io.quarkiverse.operatorsdk.annotations.CSVMetadata} instead
+ */
 public interface SharedCSVMetadata {
 }
