+++ samples/pingpong/src/main/java/io/quarkiverse/operatorsdk/samples/pingpong/PingPongOperatorCSVMetadata.java
@@ -1,7 +1,7 @@
 package io.quarkiverse.operatorsdk.samples.pingpong;
 
-import io.quarkiverse.operatorsdk.bundle.runtime.CSVMetadata;
-import io.quarkiverse.operatorsdk.bundle.runtime.SharedCSVMetadata;
+import io.quarkiverse.operatorsdk.annotations.CSVMetadata;
+import io.quarkiverse.operatorsdk.annotations.SharedCSVMetadata;
 
 @CSVMetadata(name = "pingpong-operator")
 public class PingPongOperatorCSVMetadata implements SharedCSVMetadata {
